+++ libraries/JSONPath/src/JSONPathException.php
@@ -6,12 +6,11 @@
  * @license https://github.com/SoftCreatR/JSONPath/blob/main/LICENSE  MIT License
  */
 
-declare(strict_types=1);
-
 namespace Flow\JSONPath;
 
 use Exception;
 
 class JSONPathException extends Exception
 {
+    // does nothing
 }