+++ libraries/JSONPath/CHANGELOG.md
@@ -1,11 +1,26 @@
 # Changelog
 
+### 0.9.0
+🔻 Breaking changes ahead:
+
+- Dropped support for PHP < 8.1
+
+### 0.8.3
+- Change `getData()` so that it can be mixed instead of array
+
+### 0.8.2
+- AccessHelper & RecursiveFilter now return a plain `object`, rather than an `ArrayAccess` object
+
+### 0.8.1
+- Removed strict_types
+- Applied some PSR-12 related changes
+- Small code optimizations
+
 ### 0.8.0
 🔻 Breaking changes ahead:
 
  - Dropped support for PHP < 8.0
  - Removed deprecated method `JSONPath->data()`
-
 
 ### 0.7.5
  - Added support for $.length